Rename credential and signedCredential modules files names
