+++ ts/credentials/signedCredential/signedCredential.ts
