+++ ts/identityWallet/identityWallet.ts
@@ -1,4 +1,7 @@
 import { DidDocument } from '../identity/didDocument';
+import { Credential } from '../credentials/credential/credential';
+import { SignedCredential } from '../credentials/signedCredential/signedCredential';
+import { CredentialRequest } from '../credentialRequest';
 
 export class IdentityWallet {
   private privateIdentityKey: Buffer
@@ -9,6 +12,15 @@ export class IdentityWallet {
     const identityWallet = new IdentityWallet()
     identityWallet.privateIdentityKey = privateIdentityKey
     return identityWallet
+  }
+
+  public create = {
+    credential: Credential.create,
+    signedCredential: (credentialAttrs) => SignedCredential.create({
+      credentialAttrs,
+      privateIdentityKey: this.privateIdentityKey
+    }),
+    credentialRequest: CredentialRequest // include the method after credential request PR is merged
   }
 
   // TODO: change to be an instance of Identity class
